FEAT: 동행복권 API를 사용하여 회차 정보 및 회차 별 당첨 금액 가져오기 구현 완료

+++ src/main/java/com/icebear2n2/lotto/model/dto/DhlotteryRoundDto.java
@@ -9,13 +9,13 @@ import java.util.Date;
 @Getter
 public class DhlotteryRoundDto {
     @JsonProperty("totSellamnt")
-    private BigDecimal totSellamnt;
+    private Long totSellamnt;
     @JsonProperty("returnValue")
     private String returnValue;
     @JsonProperty("drwNoDate")
     private Date drwNoDate;
     @JsonProperty("firstWinamnt")
-    private BigDecimal firstWinamnt;
+    private Long firstWinamnt;
     @JsonProperty("drwtNo6")
     private Integer drwtNo6;
     @JsonProperty("drwtNo4")
@@ -27,7 +27,7 @@ public class DhlotteryRoundDto {
     @JsonProperty("bnusNo")
     private Integer bnusNo;
     @JsonProperty("firstAccumamnt")
-    private BigDecimal firstAccumamnt;
+    private Long firstAccumamnt;
     @JsonProperty("drwNo")
     private Long drwNo;
     @JsonProperty("drwtNo2")
